LPS-171617 Move logic to a common base class and validate the existence of previous Configuration over both configurations to delete them if it existed previously
